requirements.txt
grpcio>=1.47.0
protobuf>=1.20
geodesic-api>=0.6.0
torch==1.8.1
torchvision==0.9.1
fastai==1.0.61
scikit-learn==1.0.1
timm==0.6.7
arcgis==2.0.1
